requirements.txt
# Web framework
fastapi>=0.111
uvicorn[standard]>=0.29
gunicorn>=22.0

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13

# Templates
jinja2>=3.1

# Security
cryptography>=42.0
python-jose[cryptography]>=3.3
bcrypt>=4.0
itsdangerous>=2.1

# Forms & sessions
python-multipart>=0.0.9
starlette>=0.37

# QR codes
qrcode[pil]>=7.4

# Excel export
openpyxl>=3.1

# Rate limiting
slowapi>=0.1.9

# Settings
pydantic-settings>=2.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
